+++ lib/extrapermissions.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2022
+ * (c) Copyright Ascensio System SIA 2023
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -295,7 +295,11 @@ class ExtraPermissions {
 
         $value = is_array($values) ? $values : [];
 
-        return $value;
+        return [
+            "id" => (int)$value["id"],
+            "share_id" => (string)$value["share_id"],
+            "permissions" => (int)$value["permissions"]
+        ];
     }
 
     /**
@@ -325,9 +329,18 @@ class ExtraPermissions {
 
         $values = $result ? $select->fetchAll() : [];
 
-        $value = is_array($values) ? $values : [];
+        $result = [];
+        if (is_array($values)) {
+            foreach ($values as $value) {
+                array_push($result, [
+                    "id" => (int)$value["id"],
+                    "share_id" => (string)$value["share_id"],
+                    "permissions" => (int)$value["permissions"]
+                ]);
+            }
+        }
 
-        return $value;
+        return $result;
     }
 
     /**