+++ appinfo/application.php
@@ -32,6 +32,7 @@ use OCP\Files\Template\ITemplateManager;
 use OCP\Files\Template\TemplateFileCreator;
 use OCA\Files\Event\LoadAdditionalScriptsEvent;
 use OCP\Files\IMimeTypeDetector;
+use OCP\Files\Lock\ILockManager;
 use OCP\IL10N;
 use OCP\IPreview;
 use OCP\ITagManager;
@@ -59,6 +60,7 @@ use OCA\Onlyoffice\Notifier;
 use OCA\Onlyoffice\Preview;
 use OCA\Onlyoffice\TemplateManager;
 use OCA\Onlyoffice\TemplateProvider;
+use OCA\Onlyoffice\SettingsData;
 
 use Psr\Container\ContainerInterface;
 
@@ -139,6 +141,10 @@ class Application extends App implements IBootstrap {
             );
         });
 
+        $context->registerService("SettingsData", function (ContainerInterface $c) {
+            return new SettingsData($this->appConfig);
+        });
+
         // Controllers
         $context->registerService("SettingsController", function (ContainerInterface $c) {
             return new SettingsController(
@@ -197,7 +203,8 @@ class Application extends App implements IBootstrap {
                 $this->crypt,
                 $c->get("IManager"),
                 $c->get("Session"),
-                $c->get(ITagManager::class)
+                $c->get(ITagManager::class),
+                $c->get(ILockManager::class)
             );
         });
 
@@ -212,7 +219,8 @@ class Application extends App implements IBootstrap {
                 $c->get("Logger"),
                 $this->appConfig,
                 $this->crypt,
-                $c->get("IManager")
+                $c->get("IManager"),
+                $c->get(ILockManager::class)
             );
         });